refactor fix issues with nextDateCalculator
